requirements.txt
# Core Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0
dgl>=1.1.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
networkx>=3.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Bioinformatics
biopython>=1.81

# Explainability
shap>=0.42.0
captum>=0.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
pyvis>=0.3.2

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
wandb>=0.15.0
tensorboard>=2.13.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0
